Write methods
1-ListManagers():
This method retrieves all employees who hold the position of "Manager."

2-GetReservationsByCustomer(CustomerId):
This method takes a customer identifier as a parameter and returns a list of reservations made by that particular customer.

3-ListOrdersAndMenuItems(ReservationId):
This method takes a reservation identifier as a parameter and lists the orders placed on that specific reservation along with the associated menu items.

4-ListOrderedMenuItems(ReservationId):<< I implemented this in two ways >>
This method takes a reservation identifier as a parameter and finds the menu items ordered in that specific reservation.

5-CalculateAverageOrderAmount(EmployeeId):
 This method takes an employee identifier as a parameter and calculates the average order amount for that specific employee.